Added a way to autoload extensions
This is still a WIP but it's working.
To autoload a lua extension it should contain an "autoload" function.
Autoloading doesn't trigger the "activate" function by default, it triggers
the "autoload" function.
A way to prevent Autoloading by user is missing and this could be dangerous.
The only way to prevent and extension from autoloading for the moment
is to remove the autoload function from its script.